% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwplotAddins.R
\name{dtmelt}
\alias{dtmelt}
\title{Use data.table::melt}
\usage{
dtmelt(df, ...)
}
\arguments{
\item{df}{a data.frame
dimension}

\item{...}{remaining arguments}
}
\value{
A melted data.frame
}
\description{
Use data.table::melt
}
