% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwplotAddins.R
\name{dtouter}
\alias{dtouter}
\title{Replacement for reshape on outer}
\usage{
dtouter(N)
}
\arguments{
\item{N}{dimension}
}
\value{
A data.frame corresponding to `melt` on `outer(1:N,1:N)`
}
\description{
Replacement for reshape on outer
}
